Refactor structs into a util file
